Update SSE3 check on Windows
As we no longer use Python as a proxy in `conda-build` for Visual Studio
version, this check no longer works correctly on `conda-build` 3.
Fortunately this is easy to fix and works more cleanly with
`conda-build` 3 thanks to Visual Studio environment variables that can
be inspected. So fix this to check the Visual Studio version directly
and determine whether to enable SSE3 instructions or not.