recipe/bld.bat: build both static and shared libraries
Note that the new package will have a different structure than the
current package provides a static library, `Library/lib/pixman-1.lib`.
The new package will provide the static library in
`Library/lib/libpixman-1.a`, a DLL in `Library/bin/libpixman-1-0.dll`,
and a DLL stub library in `Library/lib/pixman-1.lib`. So the semantics
of the `pixman-1.lib` file will have changed slightly, but in a way that
aligns with modern usage.